+++ incl-master/incl-linux32/malloc.h
@@ -5,10 +5,10 @@
 
 extern "C" {
     #ifndef __GNUC__
-	void *malloc(unsigned);
+	void *malloc(size_t);
 	void free(void *);
-	void *realloc(void *, unsigned);
-	void *calloc(unsigned, unsigned);
+	void *realloc(void *, size_t);
+	void *calloc(size_t, size_t);
     #else
 	#include <stdlib.h>
     #endif
