+++ src/size.C
@@ -69,9 +69,9 @@ int get_line(FILE* fp)
 	if (strcmp("DLARGEST_INT",s) == 0) { 
 		if (fscanf(fp," %s",s2)==EOF) return 0;
 //// get rid of quotes
-		for(int i=0;i<=strlen(s2)+1;i++)
+		for(size_t i=0;i<=strlen(s2)+1;i++)
 			if (s2[i]== '"') {
-				for (int j=i;j<=strlen(s2)+1;j++)
+				for (size_t j=i;j<=strlen(s2)+1;j++)
 					s2[j]=s2[j+1];
 			}
 		}
